refactor: switch to wildcards importing files when multi-targetting
