+++ src/Xbehave.Net35/packages.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="ILMerge" version="2.13.0307" targetFramework="net35" developmentDependency="true" />
-  <package id="LiteGuard.Source" version="0.9.0" targetFramework="net35" developmentDependency="true" />
   <package id="StyleCop.MSBuild" version="4.7.49.0" targetFramework="net35" developmentDependency="true" />
   <package id="xunit" version="1.9.2" targetFramework="net35" />
   <package id="xunit.extensions" version="1.9.2" targetFramework="net35" />
