+++ src/Xbehave.Net40/Xbehave.Net40.csproj
@@ -74,6 +74,7 @@
     <Compile Include="StepContext.cs" />
     <Compile Include="StepExtensions.cs" />
     <Compile Include="StepType.cs" />
+    <Compile Include="StringExtensions.Async.cs" />
     <Compile Include="StringExtensions.cs" />
     <Compile Include="_.cs" />
     <None Include="Sandbox\DynamicStringExtensions.cs" />
