+++ src/Xbehave.Sdk.Net40/AsyncStep.cs
@@ -2,6 +2,7 @@
 //  Copyright (c) xBehave.net contributors. All rights reserved.
 // </copyright>
 
+#if NET45
 namespace Xbehave.Sdk
 {
     using System;
@@ -48,3 +49,4 @@ namespace Xbehave.Sdk
         }
     }
 }
+#endif
