+++ src/test/Xbehave.Test.Acceptance.Net40/AsyncStepFeature.cs
@@ -2,6 +2,7 @@
 //  Copyright (c) xBehave.net contributors. All rights reserved.
 // </copyright>
 
+#if NET45
 namespace Xbehave.Test.Acceptance
 {
     using System;
@@ -273,3 +274,4 @@ namespace Xbehave.Test.Acceptance
 #endif
     }
 }
+#endif
