+++ src/test/Xbehave.Test.Acceptance.Net40/Infrastructure/CustomEqualityComparer.cs
@@ -2,6 +2,7 @@
 //  Copyright (c) xBehave.net contributors. All rights reserved.
 // </copyright>
 
+#if !V2
 namespace Xbehave.Test.Acceptance.Infrastructure
 {
     using System;
@@ -29,3 +30,4 @@ namespace Xbehave.Test.Acceptance.Infrastructure
         }
     }
 }
+#endif
