+++ src/test/Xbehave.Test.Acceptance.Net40/StepTimeoutFeature.cs
@@ -2,6 +2,7 @@
 //  Copyright (c) xBehave.net contributors. All rights reserved.
 // </copyright>
 
+#if !V2
 namespace Xbehave.Test.Acceptance
 {
     using System;
@@ -94,3 +95,4 @@ namespace Xbehave.Test.Acceptance
 #endif
     }
 }
+#endif
